Déploiement_Streamlit/requirements.txt
Flask
flask-cors
numpy
Pillow
python-dotenv
google-cloud-storage
absl-py
astunparse
blinker
certifi
charset-normalizer
click
flatbuffers
gast
google-pasta
grpcio
gunicorn
h5py
idna
itsdangerous
Jinja2
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
opt_einsum
optree
packaging
pip
protobuf
Pygments
requests
rich
setuptools
six
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
typing_extensions
urllib3
Werkzeug
wheel
wrapt
pytest
streamlit